+++ QuizClear/src/main/java/com/uth/quizclear/repository/CLORepository.java
@@ -1,6 +1,6 @@
 package com.uth.quizclear.repository;
 
-import com.uth.quizclear.model.CLO;
+import com.uth.quizclear.model.entity.CLO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.domain.Page;
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CLORepository extends JpaRepository<CLO, Integer> {
+public interface CLORepository extends JpaRepository<CLO, Long> {
     // Sửa từ findByCloName thành một trong những option sau:
     
     // Option 1: Tìm theo cloCode (thường dùng nhất)
