+++ QuizClear/src/main/java/com/uth/quizclear/repository/UserRepository.java
@@ -1,12 +1,12 @@
 package com.uth.quizclear.repository;
 
-import com.uth.quizclear.model.User;
+import com.uth.quizclear.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
